Fixed the size of the terminal in the browser.
Finally, we got a full-width and full-height terminal in the browser page (specifically in the DOM element) which works smoothly. Before, it was a hack to make it *almost* full width & height, but now the experience is much, much cooler!

Big thanks to @lcristianiim for letting me know how to do that! :cake: